refactor: extract WebpackBuilder
