+++ packages/builder/src/index.js
@@ -1,1 +1,6 @@
-export { default as Builder } from './builder'
+import BasicBuilder from './builders/basic'
+import WebpackBuilder from './builders/webpack'
+
+export { BasicBuilder }
+export { WebpackBuilder }
+export { WebpackBuilder as Builder }
